Add : 아이템 코멘트 값이 없을때 빈 컴포넌트 추가
