+++ src/pages/pandaitem/PandaItemCommnetTextArea.jsx
@@ -1,0 +1,15 @@
+
+export default function PandaItemCommnetTextArea() {
+  return (
+    <div className="pandaitem_comments_text_area">
+      <p className="title">문의하기</p>
+      <textarea
+        className="textarea"
+        placeholder="개인정보를 공유 및 요청하거나, 명예 훼손, 무단 광고, 불법 정보 유포시 모니터링 후 삭제될 수 있으며, 이에 대한 민형사상 책임은 게시자에게 있습니다."
+      ></textarea>
+      <div>
+        <button>등록</button>
+      </div>
+    </div>
+  );
+}
